Dimensions: cloned world in 6 scaled dimensions with entry gating (#7)
* Fix bugs and performance issues from code review

Bugs:
- Inventory.SendSlot: fix broken bounds check (&& → ||)
- NPC.DropItems: fix cast-before-null-check, use intermediate variable
- Inventory.RemoveItem: use CloneWithoutId() to preserve custom item data on stack split
- Packets.MoveCharacter: remove spurious double unary + typo

Performance:
- GameServer.GameLoop: pre-allocate read/write lists, reuse each tick
- GameServer: single shared receive buffer instead of per-read allocation
- Player: use StringBuilder for receive buffer instead of string concatenation
- GameWorld.ParseData: use IndexOf loop instead of Split to avoid array allocation
- EventHandler: cache event factory delegates to avoid Activator.CreateInstance per packet

* Add dimensions design doc

* Add dimensions part 1 implementation plan

* Add dimensions part 2 implementation plan

* docs: revise dimensions plans after review feedback

Part 1 gains two server extension points that Part 2 would otherwise have
reached around and got subtly wrong:

- NPCHandler.AddNPC/SpawnNPC, because LoadFromTemplate leaves NPCs out of
  NPCHandler.npcs and therefore out of NPCCount. LoadNPCs is refactored onto
  it so there is one definition of "spawned".
- Map.CloneAs, because requiredItems is private and a script-built clone
  silently drops item-gated entry. Muted was being missed too.

The long widening no longer leans on compiler errors - explicit narrowing
casts compile fine and truncate (AttributeSet.operator* casts HP/MP to int
today). Replaced with an 18-row audit list plus tests that multiply and copy
an AttributeSet past int.MaxValue and run a high-damage template through a
real NPC's damage path.

CanPlayerJoin now fails closed: a gate script that throws refuses entry
rather than admitting the player. Added end-to-end tests through
Map.PlayerCanJoin covering the refusal, the GM bypass and the throw.

Part 2:

- The unlock chain is written out in full, including DimensionUnlock.csx,
  which was referenced but never specified. Wardens are built from
  configuration - appearance, class, level - are invincible, and skip the
  scaler. Warden class is validated up front because NPC.LoadFromTemplate
  dereferences Class.GetLevel(Level) with no null check and class 1 has no
  level-50 row.
- New ally-rewiring pass. Ally checks compare template references, so
  dimension mobs allied to dimension-0 templates recognised nothing.
- Bound maps are clamped alongside the current map, or a player whose
  progress is reduced keeps a locked-dimension bind and returns by dying.
- The test fixture installs all three shipped scripts and creates
  Scripts/Quest, so the cloning tests can actually run.
- Collision preflight for every generated id space; AddTemplate and AddQuest
  both overwrite silently.
- The disabled-flag path keeps a test after the flag flips, instead of
  losing the only check that Enabled=false restores stock behaviour.

The boss spawn stays data: BossTemplateId names a template spawned in
dimension 0 and CloneSpawns copies it outward. The script never creates
spawns.

Test counts corrected against a measured baseline of 229 passed
(Goose.Tests 105, Tools.Tests 124): Part 1 adds 32 cases, not 14, with a
per-task running total.

* feat: add idempotent SQLite column migration on startup

* feat: add PropertiesDictionary for typed script property storage

* feat: persist Player.Properties in players.player_properties

* fix: snapshot player save scalars on the game thread

* feat: widen HP, MP and WeaponDamage to long for dimension scaling

Abyss scaling formulas overflow int.MaxValue from dimension 3 (boss-tier)
and dimension 5 (everything). Widen AttributeSet.HP/MP, all WeaponDamage
paths, and the DB reads that feed them to long. SP stays int.

Audit list outcomes (all 18 rows):
  1  AttributeSet.HP/.MP               int -> long. DONE
  2  operator+ / operator- (HP,MP)     now long arithmetic; no cast. VERIFIED
  3  operator* HP/MP (:180,181)        (int)Math.Ceiling -> (long)Math.Ceiling. DONE
  4  Clone (:75,76)                    copies long -> long; not narrowing. VERIFIED
  5  NPCTemplate.WeaponDamage          int -> long. DONE
  6  NPC.WeaponDamage (:302)           int -> long. DONE
  7  ICharacter.WeaponDamage (:88)     int -> long. DONE
  8  Player.WeaponDamage (:1572) +     int -> long (return widens); item damage
     Inventory.GetWeaponDamage         stays int (items not dimension-scaled). DONE
  9  Pet.WeaponDamage (Pet.cs:76)      override int -> long. DONE
 10  Pet DB read (Pet.cs:265)          Convert.ToInt32 -> Convert.ToInt64. DONE
 11  Pet INSERT/UPDATE (:361,:432)     string concat; unchanged, long formats
                                       identically. VERIFIED
 12  NPCHandler DB reads (:54,81,82)   Convert.ToInt32 -> Convert.ToInt64. DONE
 13  Player DB reads player_hp/mp      Convert.ToInt32 -> Convert.ToInt64. DONE
 14  NPC.Attack damage (NPC.cs:1370)   double accumulator; long -> double implicit,
                                       unchanged. VERIFIED
 15  Player.Attack (1543-1549)         double accumulator; unchanged; WeaponDamage
                                       == 1 sentinel still reads correctly (long vs
                                       int literal 1). VERIFIED
 16  SpellEffect formula symbols       Dictionary<string, decimal>; long -> decimal
      (:1274,:1284)                    implicit, unchanged. VERIFIED
 17  Packets with HP/MP/damage         string concat; verified no int.Parse
                                       round-trip. Packets.cs:102/183/211/409/410
                                       cast only the 0-100 HP% ratio, not raw HP.
                                       Unchanged. VERIFIED
 18  Window.cs:285 pet weapon damage   string concat; unchanged. VERIFIED

Additional Convert.ToInt32 -> ToInt64 sites caught by the plan's grep
`Convert.ToInt32 | grep -iE hp|mp|weapon_damage` (same columns/fields feeding
the widened values; ToInt32 would throw OverflowException past int.MaxValue):
ClassHandler.cs:85,86 (class player_hp/mp), SpellHandler.cs:87,88 (spell
effect hp/mp), Pet.cs:230,231 (pet_hp/mp), ItemHandler.cs:77,78 (item
template player_hp/mp).

Left int on purpose (not dimension-scaled / not HP/MP/WeaponDamage values):
ItemTemplate.WeaponDamage, Item.WeaponDamage/TotalWeaponDamage,
IItem.WeaponDamage, regen fields, spell HP/MP costs, HP% packet ratio casts.

* feat: allow scripts to register and copy NPC templates and spawn NPCs

* fix: route /spawnnpc through NPCHandler.SpawnNPC

* feat: add Map.CloneAs for script-generated map variants

* feat: expose QuestHandler to scripts with AddQuest

* fix: initialize Quest.PrerequisiteQuests so script quests cannot crash the quest window

* feat: add IMapScript.CanPlayerJoin entry gate

* feat: raise MaxNPCs to 250000 for dimension NPC population

* fix: stop narrowing long quest HP/MP rewards to int

* test: add fixture for compiling and running global scripts

* feat: add Dimensions.csx skeleton with configuration

* feat: clone and scale NPC templates per dimension

* test: keep the disabled-flag path covered after enabling dimensions

* feat: rewire dimension NPC allies to same-dimension templates

* feat: clone maps per dimension and rewire warps in-dimension

* feat: clone NPC spawns onto dimension maps

* feat: add DimensionMap.csx entry gating and login clamp

* fix: honor GM map-requirements bypass in DimensionMap login gate

* feat: add /dimension warp command

* feat: add warden NPCs and the dimension unlock quest chain

* docs: record dimension world load measurements

+++ Goose/ClassHandler.cs
@@ -82,8 +82,8 @@ namespace Goose
                 c.Experience = Convert.ToInt64(reader["level_up_exp"]);
 
                 c.BaseStats = new AttributeSet();
-                c.BaseStats.HP = Convert.ToInt32(reader["player_hp"]);
-                c.BaseStats.MP = Convert.ToInt32(reader["player_mp"]);
+                c.BaseStats.HP = Convert.ToInt64(reader["player_hp"]);
+                c.BaseStats.MP = Convert.ToInt64(reader["player_mp"]);
                 c.BaseStats.SP = Convert.ToInt32(reader["player_sp"]);
                 c.BaseStats.AC = Convert.ToInt32(reader["stat_ac"]);
                 c.BaseStats.Strength = Convert.ToInt32(reader["stat_str"]);
