+++ Goose/ItemHandler.cs
@@ -74,8 +74,8 @@ namespace Goose
                     template.Description = Convert.ToString(reader["item_description"]);
 
                     template.BaseStats = new AttributeSet();
-                    template.BaseStats.HP = Convert.ToInt32(reader["player_hp"]);
-                    template.BaseStats.MP = Convert.ToInt32(reader["player_mp"]);
+                    template.BaseStats.HP = Convert.ToInt64(reader["player_hp"]);
+                    template.BaseStats.MP = Convert.ToInt64(reader["player_mp"]);
                     template.BaseStats.SP = Convert.ToInt32(reader["player_sp"]);
                     template.BaseStats.AC = Convert.ToInt32(reader["stat_ac"]);
                     template.BaseStats.Strength = Convert.ToInt32(reader["stat_str"]);
